Moving network manager header file to private
