+++ demos/common/ota/aws_iot_ota_update_demo.c
@@ -61,7 +61,7 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 /* Amazon FreeRTOS OTA agent includes. */
 #include "aws_ota_agent.h"
 
-#include "iot_network_manager.h"
+#include "iot_network_manager_private.h"
 /* Required for demo task stack and priority */
 #include "aws_ota_update_demo.h"
 #include "aws_demo_config.h"
