+++ demos/espressif/esp32_devkitc_esp_wrover_kit/common/application_code/main.c
@@ -56,7 +56,7 @@
 #include "semphr.h"
 #include "driver/uart.h"
 #include "aws_application_version.h"
-#include "iot_network_manager.h"
+#include "iot_network_manager_private.h"
 
 
 #if BLE_ENABLED
