+++ demos/nordic/nrf52840-dk/common/application_code/main.c
@@ -62,7 +62,7 @@
 
 #include "iot_ble.h"
 #include "iot_ble_numericComparison.h"
-#include "aws_iot_network_manager.h"
+#include "iot_network_manager_private.h"
 #include "SEGGER_RTT.h"
 #include "aws_application_version.h"
 #include "iot_taskpool.h"
