+++ lib/include/private/iot_network_manager_private.h
